+++ src/plonemeeting/portal/core/content/institution.py
@@ -240,8 +240,9 @@ class IInstitution(model.Schema):
 class Institution(Container):
     """
     """
-    def get_delib_categories(self):
+    def fetch_delib_categories(self):
         categories = []
+        print("fetch_delib_categories")
         if self.plonemeeting_url and self.meeting_config_id and self.username and self.password:
             delib_config_category_field = CATEGORY_IA_DELIB_FIELDS_MAPPING_EXTRA_INCLUDE[self.delib_category_field]
             url = "{plonemeeting_url}/@config?config_id={meeting_config_id}&extra_include={delib_category_field}".format(
@@ -257,5 +258,5 @@ class Institution(Container):
 
             for cat in cat_json:
                 categories.append((cat['id'], cat['title']))
-
+            self.delib_categories = categories
         return categories